moved default tokens to constants
